Comments to the visualizations in the notebook
Commented on the predictive analysis, as well as the median, mean, skewness, and standard deviation. Commented on the histograms based on temperature and humidity. In addition, commented on the weather data collected during the period for the different cities. Also fixed a small issue with data storage in src/main.